feat: Add groupBy utility function
Adds a `groupBy` utility function that creates an object composed of keys generated from the results of running each element of a collection thru an iteratee.

This function is useful for grouping a collection of objects by a common property.

Includes a test suite covering grouping by a function and by a property name, as well as handling empty and non-array inputs.