feat(utils): add once utility
Implements a `once` function that creates a version of a function that can only be called one time. Subsequent calls to the function will return the result of the first call.

This is useful for initialization functions or any logic that should only run a single time.

Includes unit tests covering basic functionality, argument passing, and `this` context.